polynomial: Copy existing polynomial implementation to new crate.
Other than a few working and import changes to reflect the new
location of this module/crate, this is a verbatim copy of the
implementation of polynomials that currently exists in the
swanky-field crate.

With this crate having a complete initial implementation (refactoring,
documenting, and testing will be the subjets of future work), the old
implementation can be incrementally replaced.