field_test: Remove proptests that depend on polynomial_modulus.
The tests polynomial_mul and lifted_polynomial_mul depend on the
FiniteField method polynomial_modulus, which will be removed in a
future commit.

These tests do test a meaningful/interesting property, but only for
fields with a non-trivial polynomial_modulus. These tests will be
restored for such fields in a future commit.