field: Remove the polynomial_modulus method from FiniteField.
This commit removes the method polynomial_modulus from FiniteField,
and updates all implementations of the trait provided by Swanky to
match.

This change was motivated by the fact that, in general, a finite field
does not have a meaningful notion of "polynomial modulus". This
concept is only relevant for non-prime (prime-power) finite fields,
the elements of which are polynomials reduced over some a priori
chosen irreducible polynomial of appropriate degree.

Indeed, the only uses of the polynomial_modulus method were the tests
removed in the previous commit; as stated there, those tests are
meaningful, but only for some fields -- those described here, which
are non-prime.

For such fields, this commit also lifts the implementation of
polynomial_modulus to the module level, exposing them as pub and
renaming to specify which field's polynomial modulus is returned. When
the previously-removed tests are re-instated (i.e. via macro), these
moduli will be used.