swanky-polynomial: Update downstream test code.
In a previous commit, all non-test code was updated to use the new
polynomial crate, but tests were left untouched.

This commit completes that work by updating all downstream test code
involving the use of polynomials to use the new crate. In some cases,
this involved introducing an additional dependency on the new crate to
allow the test-generating macro to work. In the future, these
dependencies may be removed when polynomials are no longer tested in
the field-testing crate.