field crates: Make swanky-polynomial a dev dependency when possible.
Previously, swanky-polynomial was a regular dependency of the
swanky-field-binary and swanky-field-ff-primes crates, when
polynomials are only used in testing these fields.

This commit updates the visibility of the recently-added
polynomial_modulus_** functions to private and configures them to be
test-only, meaning the import of swanky-polynomial could also be made
test-only, thus making it possible to make swanky-polynomial a dev
dependency.