vectoreyes/build/backend/avx2: Fix Boolean handling.
The comment included in this commit documents the situation, but for
reviewers/posterity: quick-xml 0.37 updated the handling of Booleans
to only support the strings considered valid in the Xml Schema. Since
the Intel Intrinsics XML file uses an invalid string for Boolean
values, this commit adds a handler that accepts anything the previous
version of quick-xml accepted for our Booleans.