Cargo.toml: Upgrade pinned crates.
IMPORTANT NOTE: This commit sets ndarray to be the specific version
0.15.6, since issue #123 has finally caught up from the other end:
Humidor / Ligero break with the latest version of ndarray, in a
similar way to the bug described in that ticket.

Long-term, since ndarray is required by more than Humidor (in
particular, the prime fields from the ff crate), we probably don't
want to stick with this solution. I'm not sure we have the time/budget
to invest in fixing this more legitimately.

In addition to this commit message, there will be a note added to
issue #123 pointing to the relevant MR / describing the issue in
greater detail.